Fix/refactor general

+++ src/main/java/com/elasbancam/repositories/PessoaJuridicaRepository.java
@@ -1,10 +1,9 @@
 package com.elasbancam.repositories;
 
-import com.elasbancam.models.PessoaFisica;
 import com.elasbancam.models.PessoaJuridica;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Integer> {
+public interface PessoaJuridicaRepository extends JpaRepository<PessoaJuridica, Long> {
 }
